CraftPlayer was in the wrong place... rip
