+++ src/main/java/org/bukkit/craftbukkit/entity/CraftPlayer.java
@@ -20,9 +20,14 @@ import java.util.logging.Logger;
 
 import com.mojang.authlib.GameProfile;
 
+import net.md_5.bungee.api.chat.BaseComponent;
 import net.minecraft.entity.EntityTracker;
 import net.minecraft.entity.EntityTrackerEntry;
 import net.minecraft.entity.player.EntityPlayerMP;
+import net.minecraft.network.Packet;
+import net.minecraft.network.play.client.C01PacketChatMessage;
+import net.minecraft.network.play.server.S02PacketChat;
+import net.minecraft.server.MinecraftServer;
 import net.minecraft.network.play.server.S23PacketBlockChange;
 import net.minecraft.network.play.server.S33PacketUpdateSign;
 import net.minecraft.network.play.server.S38PacketPlayerListItem;
@@ -1339,6 +1344,22 @@ public class CraftPlayer extends CraftHumanEntity implements Player {
         {
            return getHandle().translator;
         }
+
+        @Override
+        public void sendMessage(BaseComponent component) {
+            this.sendMessage(new BaseComponent[]{component});
+        }
+
+        @Override
+        public /* varargs */ void sendMessage(BaseComponent ... components) {
+            if (CraftPlayer.this.getHandle().playerNetServerHandler == null) {
+                return;
+            }
+            S02PacketChat packet = new S02PacketChat();
+            packet.components = components;
+            CraftPlayer.this.getHandle().playerNetServerHandler.sendPacket(packet);
+        }
+
     };
 
     public Player.Spigot spigot()
